Update Readme.md to remove ambiguity for personal token
There has been recent change in the ways to create GitHub API token. Earlier there was only one option to create token but now with the recent change one more option is added i.e. Generate new token (Fine-grained, repo-scoped). 
The current documentation assumes we are creating a token with earlier option i.e. Generate new token (For general use) and thus says "You'll need a GitHub API Token with repo and user scope...". With the advent of new option this is invalid. 
Hence we need to specify in the Readme.md file to create the token using previous option only.
This commit updates that.